move files to mobile dir
